chore: change all int32 values to int64 to avoid api issues

+++ outscale/resource_outscale_nic_link.go
@@ -70,7 +70,7 @@ func resourceOutscaleOAPINetworkInterfaceAttachmentCreate(d *schema.ResourceData
 	nicID := d.Get("nic_id").(string)
 
 	opts := oscgo.LinkNicRequest{
-		DeviceNumber: int32(di),
+		DeviceNumber: int64(di),
 		VmId:         iID,
 		NicId:        nicID,
 	}
